+++ dags/src/yt_video.py
@@ -1,7 +1,7 @@
 import os
 import json
 
-from src.yt_utils import logger, youtube, drop_location
+from src.yt_utils import logger, youtube, LOCAL_DATA_DIR
 
 
 def main(video_ids):
@@ -26,11 +26,15 @@ def main(video_ids):
                 videos_data.append(
                     {
                         "videoId": item["id"],
+                        "title": item["snippet"]["title"],
                         "channelId": item["snippet"]["channelId"],
                         "categoryId": item["snippet"]["categoryId"],
                         "defaultAudioLanguage": item["snippet"].get(
                             "defaultAudioLanguage", ""
                         ),
+                        "videoPublishedAt": item["snippet"]["publishedAt"],
+                        "description": item["snippet"]["description"],
+                        "defaultLanguage": item["snippet"].get("defaultLanguage", ""),
                         "tags": item["snippet"].get("tags", []),
                         "duration": item["contentDetails"]["duration"],
                         "definition": item["contentDetails"]["definition"],
@@ -51,14 +55,16 @@ def main(video_ids):
 
     json_response = json.dumps(videos_data, indent=2)
 
-    if not os.path.exists(f"{drop_location}/{channel_id}"):
+    if not os.path.exists(f"{LOCAL_DATA_DIR}/{channel_id}"):
         # Create the directory if it does not exist
-        logger.info(f"Creating directory {drop_location}")
-        os.mkdir(f"{drop_location}/{channel_id}")
+        logger.info(f"Creating directory {LOCAL_DATA_DIR}")
+        os.mkdir(f"{LOCAL_DATA_DIR}/{channel_id}")
 
     # Write the response to a file
-    open(f"{drop_location}/{channel_id}/video_info.json", "w").write(str(json_response))
-    logger.info(f"Response written to {drop_location}/{channel_id}.json")
+    open(f"{LOCAL_DATA_DIR}/{channel_id}/video_info.json", "w").write(
+        str(json_response)
+    )
+    logger.info(f"Response written to {LOCAL_DATA_DIR}/{channel_id}.json")
 
 
 if __name__ == "__main__":